WIP on refactoring dialogs
